Added some Uint32Math overflow checks in JSON strings. Change GrowingArray to use uint32 and added some overflow checks Changed SparseArrays to use uint32 for index and check for overflow Sparse Array refactor Removed unused function in Buffer Builder Fix for security review 5120216: CompoundString alignment Fix for security review 5120248: SparseArray Allocation Code review fixes

+++ lib/Backend/GlobOpt.cpp
@@ -441,7 +441,7 @@ GlobOpt::ForwardPass()
     // functions with constants.  There will be a gap in the symbol numbering between the main constants and
     // the inlined ones, so we'll most likely need a new array chunk.  Make the min size of the array chunks be 64
     // in case we have a main function with very few constants and a bunch of constants from inlined functions.
-    this->byteCodeConstantValueArray = SparseArray<Value>::New(this->alloc, max((int)this->func->GetJnFunction()->GetConstantCount(), 64));
+    this->byteCodeConstantValueArray = SparseArray<Value>::New(this->alloc, max(this->func->GetJnFunction()->GetConstantCount(), 64U));
     this->byteCodeConstantValueNumbersBv = JitAnew(this->alloc, BVSparse<JitArenaAllocator>, this->alloc);
     this->tempBv = JitAnew(this->alloc, BVSparse<JitArenaAllocator>, this->alloc);
     this->prePassCopyPropSym = JitAnew(this->alloc, BVSparse<JitArenaAllocator>, this->alloc);
